requirements.txt
sqlmodel==0.0.22
fastapi==0.115.8
uvicorn==0.34.0
alembic==1.14.1
PyJWT==2.10.1
bcrypt==4.2.1
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
email_validator==2.2.0
cloudpickle==3.1.1
pandas==2.2.3
shap==0.46.0
lightgbm==4.6.0
category_encoders==2.8.0
